Use non-wrappable cell text editor
If happened quite often that the comment columns would hold multi-line
comments due to the enter key starting a new line instad of ending the
edit. (This seems to apply only for Windows, not for Linux.) So now use
the GridCellTextEditor instead of the GridCellAutoWrapStringEditor to
hopefully prevent this from happening.